test: Run custom codec on integration tests
